Added dynamic service registration for interfaces
Added implementation for automatically registering services "tagged" with a service lifetime interface in the application layer. This makes us avoid having to manually make sure that all services are registered. ex: IProductService : ITransientService will automatically be registered with a transient lifetime in the service container and be available for dependency injection