Fixed issue when deleting categories
Fixed an issue where a category could be deleted if one or more products were still connected to the category. Now the method will check if any products in the Products table got the category id matching the one for deletion. If any product exist, a Custom Exception is thrown.